Backend kodları tamamlandı

+++ Sprig Boot/src/main/java/com/example/demo/repository/EmanetRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.entity.Emanet;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmanetRepository extends JpaRepository<Emanet,Long> {
+public interface EmanetRepository extends JpaRepository<Emanet,Integer> {
 }
