+++ Sprig Boot/src/main/java/com/example/demo/repository/GorevliRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface GorevliRepository extends JpaRepository<Gorevli,Long> {
+public interface GorevliRepository extends JpaRepository<Gorevli,Integer> {
     Optional<Gorevli> findByEposta(String email);
 }
